cov-buildless: do not install csdiff from COPR
It does not work in Konflux:
```
+ dnf copr enable copr.devel.redhat.com/@osh/csdiff -y
Updating Subscription Management repositories.
Unable to read consumer identity

This system is not registered with an entitlement server. You can use subscription-manager to register.

Enabling a Copr repository. Please note that this repository is not part
of the main distribution, and quality may vary.

The Fedora Project does not exercise any power over the contents of
this repository beyond the rules outlined in the Copr FAQ at
<https://docs.pagure.org/copr.copr/user_documentation.html#what-i-can-build-in-copr>,
and packages are not held to any quality or security level.

Please do not file bug reports about these packages in Fedora
Bugzilla. In case of problems, contact the owner of this repository.
Error: Failed to connect to https://copr.devel.redhat.com/coprs/@osh/csdiff/repo/epel-9/dnf.repo?arch=x86_64: Name or service not known
```